Move stringifiers into a separate class
