+++ client/src/main/java/org/spine3/validate/Validate.java
@@ -31,8 +31,8 @@ import javax.annotation.Nullable;
 import static com.google.common.base.Preconditions.checkArgument;
 import static com.google.common.base.Preconditions.checkNotNull;
 import static com.google.common.base.Preconditions.checkState;
-import static org.spine3.base.Identifiers.EMPTY_ID;
-import static org.spine3.base.Identifiers.idToString;
+import static org.spine3.base.Stringifiers.EMPTY_ID;
+import static org.spine3.base.Stringifiers.idToString;
 
 /**
  * This class provides general validation routines.
