+++ client/src/test/java/org/spine3/base/CommandsShould.java
@@ -44,8 +44,8 @@ import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertFalse;
 import static org.junit.Assert.assertTrue;
 import static org.spine3.base.Commands.getId;
-import static org.spine3.base.Identifiers.idToString;
 import static org.spine3.base.Identifiers.newUuid;
+import static org.spine3.base.Stringifiers.idToString;
 import static org.spine3.protobuf.Durations.seconds;
 import static org.spine3.protobuf.Timestamps.getCurrentTime;
 import static org.spine3.protobuf.Timestamps.minutesAgo;
