+++ server/src/main/java/org/spine3/server/command/CommandStorage.java
@@ -45,8 +45,8 @@ import static org.spine3.base.CommandStatus.ERROR;
 import static org.spine3.base.CommandStatus.RECEIVED;
 import static org.spine3.base.Commands.generateId;
 import static org.spine3.base.Commands.getId;
-import static org.spine3.base.Identifiers.EMPTY_ID;
-import static org.spine3.base.Identifiers.idToString;
+import static org.spine3.base.Stringifiers.EMPTY_ID;
+import static org.spine3.base.Stringifiers.idToString;
 import static org.spine3.protobuf.Timestamps.getCurrentTime;
 import static org.spine3.validate.Validate.checkNotDefault;
