+++ server/src/test/java/org/spine3/server/command/CommandStorageShould.java
@@ -56,8 +56,8 @@ import static org.spine3.base.CommandStatus.RECEIVED;
 import static org.spine3.base.CommandStatus.SCHEDULED;
 import static org.spine3.base.Commands.generateId;
 import static org.spine3.base.Commands.getId;
-import static org.spine3.base.Identifiers.idToString;
 import static org.spine3.base.Identifiers.newUuid;
+import static org.spine3.base.Stringifiers.idToString;
 import static org.spine3.protobuf.AnyPacker.unpack;
 import static org.spine3.protobuf.Timestamps.getCurrentTime;
 import static org.spine3.protobuf.TypeUrl.ofEnclosed;
